// src/api/types.ts
export interface Section {
	position: number;
	components: Component[];
	name: string | null;
}

export interface Instruction {
	id: number;
	position: number;
	display_text: string;
	start_time: number;
	appliance: string | null;
	end_time: number;
	temperature: number | null;
}

export type Tag =
	| { id: 64447; type: "cuisine"; name: "british"; display_name: "British" }
	| { id: 64453; type: "cuisine"; name: "italian"; display_name: "Italian" }
	| { id: 64457; type: "cuisine"; name: "mexican"; display_name: "Mexican" }
	| {
			id: 64458;
			type: "cuisine";
			name: "middle_eastern";
// ...
	  };

export interface Show {
	id: number;
	name: string;
}

export interface Credit {
	name: string | null;
	type: string;
	slug?: string;
	image_url?: string;
	id?: number;
}

export interface Brand {
	id: number;
	slug: string;
	image_url: string;
	name: string;
}

export interface Topic {
	name: string;
	slug: string;
}

export interface TotalTimeTier {
	tier: string;
	display_tier: string;
}

export interface UserRatings {
	count_positive: number;
	score: number | null;
	count_negative: number;
}

export interface Nutrition {
	updated_at?: string;
	calories?: number;
	fat?: number;
	sugar?: number;
	carbohydrates?: number;
	protein?: number;
	fiber?: number;
}

export interface Rendition {
	container: string;
	url: string;
	duration: number;
	bit_rate: number | null;
	content_type: string;
	aspect: string;
	width: number;
	minimum_bit_rate: number | null;
	name: string;
	maximum_bit_rate: number | null;
	height: number;
	poster_url: string;
	file_size: number | null;
}

export interface Compilation {
	id: number;
	name: string;
	description: string | null;
	slug: string;
	video_id: number;
	video_url: string;
	created_at: number;
	thumbnail_url: string;
	thumbnail_alt_text: string;
	approved_at: number;
	aspect_ratio: string;
	beauty_url: string | null;
	buzz_id: number | null;
	canonical_id: string;
	country: string;
	draft_status: string;
	facebook_posts: unknown[];
	is_shoppable: boolean;
	keywords: string | null;
	language: string;
	promotion: string;
	show: Show[];
}

interface Recipe {
	id: number;
	name: string;
	slug: string;
	description: string;
	canonical_id: string;
	draft_status: string;
	promotion: string;
	show: Show;
	show_id: number;
	created_at: number;
	updated_at: number;
	approved_at: number;
	aspect_ratio: string;
	beauty_url: string | null;
	brand: Brand | null;
	brand_id: number | null;
	buzz_id: number | null;
	compilations: Compilation[];
	cook_time_minutes: number | null;
	country: string;
	credits: Credit[];
	facebook_posts: unknown[];
	inspired_by_url: string | null;
	instructions: Instruction[];
	is_one_top: boolean;
	is_shoppable: boolean;
	keywords: string;
	language: string;
	num_servings: number;
	nutrition: Nutrition;
	nutrition_visibility: string;
	original_video_url: string | null;
	prep_time_minutes: number | null;
	renditions: Rendition[];
	sections: Section[];
	servings_noun_plural: string;
	servings_noun_singular: string;
	seo_title: string;
	tags: Tag[];
	thumbnail_alt_text: string;
	thumbnail_url: string;
	tips_and_ratings_enabled: boolean;
	topics: Topic[];
	total_time_minutes: number | null;
	total_time_tier: TotalTimeTier | null;
	user_ratings: UserRatings;
	video_ad_content: string | null;
	video_id: number | null;
	video_url: string | null;
	yields: string;
}

export interface RecipeSearchResponse {
	count: number;
	results: Recipe[];
}